spring security RBAC 15%

+++ src/main/java/com/jwj/russianblue/core/entity/RbEntity.java
@@ -1,9 +1,8 @@
 package com.jwj.russianblue.core.entity;
 
-import java.math.BigDecimal;
 
 public interface RbEntity extends RbModel {
 
-	public BigDecimal getId();
+	public Long getId();
 
 }
